refactor: CouponCache 삭제 및 캐시 유효 시간 변경
