+++ src/main/java/coupon/domain/coupon/repository/CouponRepositoryAdaptor.java
@@ -1,10 +1,9 @@
 package coupon.domain.coupon.repository;
 
 import coupon.domain.coupon.Coupon;
-import coupon.infra.db.CouponCache;
 import coupon.infra.db.CouponEntity;
-import coupon.infra.db.JpaCouponRepository;
-import coupon.infra.db.RedisCouponRepository;
+import coupon.infra.db.jpa.JpaCouponRepository;
+import coupon.infra.db.redis.RedisCouponRepository;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
@@ -19,20 +18,8 @@ public class CouponRepositoryAdaptor implements CouponRepository {
     @Override
     public Coupon save(Coupon coupon) {
         CouponEntity saved = jpaCouponRepository.save(toEntity(coupon));
-        redisCouponRepository.save(toCache(saved));
+        redisCouponRepository.save(saved);
         return Coupon.from(saved);
-    }
-
-    private CouponCache toCache(CouponEntity couponEntity) {
-        return new CouponCache(
-                couponEntity.getId(),
-                couponEntity.getName(),
-                couponEntity.getDiscountMoney(),
-                couponEntity.getMinOrderMoney(),
-                couponEntity.getCategory(),
-                couponEntity.getStartDate(),
-                couponEntity.getEndDate()
-        );
     }
 
     private CouponEntity toEntity(Coupon coupon) {
@@ -48,7 +35,7 @@ public class CouponRepositoryAdaptor implements CouponRepository {
 
     @Override
     public Optional<Coupon> findById(Long id) {
-        Optional<Coupon> coupon = redisCouponRepository.findById(id).map(Coupon::fromCache);
+        Optional<Coupon> coupon = redisCouponRepository.findById(id).map(Coupon::from);
         if (coupon.isEmpty()) {
             return jpaCouponRepository.findById(id).map(Coupon::from);
         }
