--- src/main/java/coupon/infra/db/RedisCouponRepository.java
@@ -1,6 +1,0 @@
-package coupon.infra.db;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface RedisCouponRepository extends CrudRepository<CouponCache, Long> {
-}
