+++ src/main/java/coupon/infra/db/jpa/JpaCouponRepository.java
@@ -1,5 +1,6 @@
-package coupon.infra.db;
+package coupon.infra.db.jpa;
 
+import coupon.infra.db.CouponEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaCouponRepository extends JpaRepository<CouponEntity, Long> {
